completed project

+++ docket_app/views.py
@@ -1,4 +1,5 @@
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.core.paginator import Paginator
 from django.http import HttpResponse
 from django.shortcuts import redirect, render
@@ -7,19 +8,22 @@ from django.urls import reverse
 from .forms import TaskForm
 from .models import TaskBuster
 
+
 def index(request):
    
     return render(request, 'index.html',{})
-
+@login_required
 def todolist(request):
     if request.method == 'POST':
         form = TaskForm(request.POST or None)
         if form.is_valid():
-            form.save()
+            instance = form.save(commit=False)
+            instance.owner = request.user
+            instance.save()
             messages.success(request,("Added Task"))
         return redirect("todo")
     else:
-        all_tasks = TaskBuster.objects.all()
+        all_tasks = TaskBuster.objects.filter(owner=request.user)
         paginator = Paginator(all_tasks,5)
         page = request.GET.get('pg')
         all_tasks = paginator.get_page(page)
@@ -28,30 +32,66 @@ def todolist(request):
         }
         
         return render(request, 'todolist.html', context)
+@login_required
 def todolist_delete(request, pk):
-    particular_task = TaskBuster.objects.get(pk=pk)
-    particular_task.delete()
-    return redirect(request.META['HTTP_REFERER'])
+    
+    try:
+        particular_task = TaskBuster.objects.get(pk=pk)
+    except:
+        return render(request, 'warning.html',{})
 
 
+    if particular_task.owner == request.user:
+        particular_task.delete()
+    else:
+        messages.error(request,("Restricted"))
+    try:
+        destination = redirect(request.META['HTTP_REFERER'])
+    except KeyError:
+        destination = redirect('todo')
+
+    return destination
+
+@login_required
 def todolist_edit(request,pk):
     if request.method == 'POST':
         particular_task = TaskBuster.objects.get(pk=pk)
+
         form = TaskForm(request.POST or None, instance = particular_task)
+        
         if form.is_valid():
             form.save()
         messages.success(request,("Task Updated"))
-    return redirect(request.META['HTTP_REFERER'])
 
+
+    try:
+        destination = redirect(request.META['HTTP_REFERER'])
+    except KeyError:
+        destination = redirect('todo')
+
+    return destination
+@login_required
 def todolist_mark_task(request,pk):
-    
-    particular_task = TaskBuster.objects.get(pk=pk)
+    try:
+        particular_task = TaskBuster.objects.get(pk=pk)
+    except:
+        return render(request, 'warning.html',{})
 
-    if particular_task.completed == True:
-        particular_task.completed = False
+    if particular_task.owner == request.user:
+
+        if particular_task.completed == True:
+            particular_task.completed = False
+        else:
+            particular_task.completed = True
+        
+        particular_task.save()
     else:
-        particular_task.completed = True
+
+        messages.error(request,("Restricted"))
     
-    particular_task.save()
-    
-    return redirect(request.META['HTTP_REFERER'])
+    try:
+        destination = redirect(request.META['HTTP_REFERER'])
+    except KeyError:
+        destination = redirect('todo')
+
+    return destination
